requirements.txt
aenum==3.1.15
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
fake-useragent==1.4.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==0.18.3
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
line-bot-sdk==3.9.0
lxml==5.1.0
MarkupSafe==2.1.5
multidict==6.0.5
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
